fix entrypoint
